Extract CLI helpers and runtime init from main.py
